+++ src/constants.ts
@@ -1,1 +1,2 @@
 export const ROLES_BUILDER_TOKEN = 'nestjs-acl-roles';
+export const OPTIONS_TOKEN = 'nestjs-acl-options';