+++ CHANGELOG.md
@@ -2,3 +2,14 @@
 
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+## [0.2.0]
+
+### Changed
+- rename all interface to remove "I" prefix
+
+### Added
+- Add a global/check options "rejectIfNoRule". If true the service will throw an exception if no acl rule was found. This allow to ensure the implementation of a rule was done.
+
+[unreleased]: https://github.com/Pop-Code/nestjs-console/compare/v0.2.0...HEAD
+[0.2.0]: https://github.com/Pop-Code/nestjs-console/compare/v0.1.2...v0.2.0